+++ .gitignore
@@ -46,3 +46,5 @@ app.*.map.json
 # Ignore local dev settings
 *.env
 meili_data/
+*.txt
+*.json
