+++ lib/components/dialog.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:fluent_ui/fluent_ui.dart';
 import 'package:gap/gap.dart';
 
 class DialogContent extends StatelessWidget {
@@ -9,36 +9,23 @@ class DialogContent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return SimpleDialog(
+    return ContentDialog(
         title: Center(
             child: Column(
           children: [Text(title), Gap(5)],
         )),
-        children: [
-          Container(
+        content: Container(
             margin: const EdgeInsets.all(20),
             child: SelectableText(
               content.replaceAll("\n", "\n\n")
             ),
           ),
-          FractionallySizedBox(
-            widthFactor: 0.8,
-            child: TextButton(
-                onPressed: () {
-                  Navigator.of(context).pop();
-                },
-                style: ButtonStyle(
-                  backgroundColor: WidgetStatePropertyAll(
-                      Theme.of(context).colorScheme.secondary),
-                  padding: WidgetStatePropertyAll(
-                      EdgeInsets.only(top: 20, bottom: 20)),
-                ),
-                child: Text(
-                  "Zurück",
-                  style: TextStyle(
-                      color: Theme.of(context).colorScheme.onSecondary),
-                )),
-          )
-        ]);
+        actions: [
+          Center(child: FilledButton(onPressed: () {
+            Navigator.of(context).pop();
+          }, child: const Text("Zurück")),)
+          
+        ],
+        );
   }
 }
