+++ pubspec.yaml
@@ -44,6 +44,7 @@ dependencies:
   animations: ^2.0.11
   sembast_web: ^2.4.1
   system_theme: ^3.1.2
+  go_router: ^14.8.1
 
 dev_dependencies:
   flutter_test:
